requirements.txt
gymnasium[box2d]>=0.29.1,<1.3.0
gymnasium[other]>=0.29.1,<1.3.0
stable-baselines3==2.8.0
PyYAML